Fixes await termination bug and avoid notifying skipped programs
Until now, the main processing loop continued whereas there were
remaining programs that could start. When the loop terminated, the run
function returned, without waiting for programs that were still running.

In addition, we do not send programStarted nor programFinished to
skipped programs, which is confusing.

In addition, we add tests for the --single-program parameter.