Adds new-project command
This command allows the developer to create a new project. A directory
with the basic items is created, including (i) a JRE to be added
to the docker image, (ii) the compi engine .jar downloaded given its
version, to be added to the docker image, (iii) a sample pipeline.xml
file and (iv) an initial Dockerfile which creates the pipeline image by
adding the JRE, the compi engine and the pipeline.xml.