Splits cli/core subprojects and removes reserved parameter name restriction
This commit splits cli and core. By separating the cli and core, a
dependency from core to cli appeared, since the pipeline validation
module was depending on cli's RunCommand to get Compi general parameter
names. In order to remove this dependency, we implemented a new cli
approach allowing the user to separate compi parameters from pipeline
specific parameters via a "--" element in the command line.