Adds runners functionality
This commit adds "runners" functionality, which allows the user to
specify how the underlying process for each task is built. Given
the task shell code, a runner is another piece of shell code, where
${task-code} and ${task-id} contain the task code and id, respectively.
In addition, any pipeline variable is resolved in the runner shell code.

The CLI run command has a new parameter (-r) to specify an XML file with
runners.

This commit also refactorizes the variable resolving as well as the loop
management.