Fix empty and skipped foreachs
This commit fixes two bugs:
- Empty foreach tasks make the pipeline to stall, because the foreach
task is expected to finish, but without iterations a finish notification
was never generated. We fix this by adding a dummy TaskRunnable for this
foreach task which finishes immediately and produces a finish
notification for the foreach task.
- Skipped foreach tasks run the value loop generatos, which is not
strictly necessary. This is specially harmful in of="command", because
the command run always, regardless of the task is skipped or not.