Resolve variables in task and runners code via environment variables
In order to make more robust variable resolution, this commit replaces
string replacement strategy for another one based on environment
variables. Now, task and runner processes run in environments where
pipeline variables are present. This makes more robust shell codes
where, for example, variables can include quotes. For example, a task
with code: echo ${parameter} will not work if parameter is defined as:
quote char is "
the echo will fail, because before we would invoke the command:
echo quote char is "
Now, we will invoke:
echo ${parameter}
where "parameter" is an environment variable containing: quote char is "
And this works as expected.
This also eases runners code, which it will be useful to copy the task
code into a file easily with: echo $task_code > script
If $task_code is complex, including quotes, they would be removed in the
previous version.

There are several modifications in Compi rules:
- Parameter names cannot include illegal characters for variables, such
as '-', start with a number, etc.
- Parameters are no longer detected inside task codes, which was very
weak. Now, pipelines must define parameters in their parameters section
and task must declare which parameters will use in the "params"
attribute, which is a comma-separated list of parameter names. If
"params" attribute is not present, the task will not see any attribute.