Refactors CompiApp
This commit does a deep refactor of CompiApp and other classes in
consequence of this refactor.
- Foreach iterations are now subclasses of Foreach in order to remove the
getParentTasks() method in CompiApp.
- The CompiApp.setResolver() method is also removed. This method was
only called by the CLI, but by adding a Proxy pattern in the CLI, we could
resolve the need to change the resolver after the CompiApp was created.
- The responsibility of creating the Pipeline was moved outside the CompiApp.
The Pipeline is created AND ALSO VALIDATED with a static factory method in
the Pipeline class.