Adds warning for undeclared parameters
This commit adds a warning when a parameter
name is found in the code of a task, but the
task has not declared it in params attribute
nor in the as attribute (in the case of foreach
tasks).

In addition, a --abort-if-warnings/-w option is
added to abort pipeline execution if a warning
is found.