Fixes showing "Exception during execution: null" in validation
When a pipeline validates, a "Exception during execution: null" is
shown. This is caused because we cannot silence stderr totally
because yacli takes a reference to System.err before we silence
it from compi dk. We change ExitTrappedException from Exception
to Error in order to escape to the catch method in yacli that
captures Exceptions and shows this "Exception during execution"
message. In addition, if an java.lang.Error is launch, it is
wrapped inside an InvocationTargetException so we fix out cath
accordyingly.