Fixes a bug when running a specific pipeline command using a params file
In a previous refactor, the RunSpecificPipelineCommand was refactored so that the
RunCommand class set the params file in the pipeline configuration and this class
used it from this configuration. However, this had an unexpected side effect: this
way the proxy variable resolver was ignored by the CompiCli class because the
params file had a value. This commit fixes this bug.