Fixes metadata properties file
This commit fixes a problem with the previous metadata properties file: special
characters of the Properties class (such as :) were being scaped in the plain text
file, thus producing invalid values to upload to CompiHub (e.g. "http\://"). This
commit fixes this problem by introducing a WriteOnlyProperties class that writes
properties without escaping special characters. This class is used only in the
CompiHubMetadataPropertiesFile.