Adds src attribute for Task elements
This commit adds an src attribute for tasks. This attribute
for <task> and <foreach> elements should contain a relative path
to an external file containing the task source code, as an alternative
to embedding the source code inside the task element.

In addition, this commit removes the Task.exec attribute, which was
virtually a clone of the Task.toExecute attribute.