Fixes src attribute parsing
This commit fixes a bug in the src attribute parsing. The bug occurred when the path
of the pipeline file was just the pipeline name because in that case the getParent()
call resulted in null. This is fixed by detecting this situation and using
new File(".") as parent when this happens.