Fixes deployment in stable versions (not -SNAPSHOTS)
Due to the creation of different assemblies such as "dist" or
"installer", the jar-with-dependencies.jar was copied to multiple
destinations and thus maven tried to deploy it several times, leading
to an error from the nexus server.

This commit fixes this by avoiding the dist and installer from
being deployed.