Adds "binded foreachs" feature
In order to improve the level of paralellism,
this feature allows to "bind" loop iterations between two foreach tasks.
Both foreachs must have the same number of iterations, and each iteration
subtask will be a dependency of the same index iteration subtask of the
dependant foreach.

Binded foreach dependencies are specified by adding "*" as a task name
prefix.

For example:

<foreach id="loop1" ...>
</foreach>

<foreach id="loop2" after="*loop1">
</foreach>

In this case the first iteration of loop2 will be started just after
the first iteration of loop1 has finished, the second iteration of
loop2 after the same of loop1, and so on.