Adds pipeline resume feature
A new commit (compi resume) is added. With resume, a pipeline which
has not been finished can be resumed by simply invoking:

compi resume -p pipeline.xml

The same configuration is used for the new execution, and it is not
possible to change or add parameters. This feature is only intended
for resuming pipelines that have aborted due to external causes
(electric failures, internet connection problems, etc.).

The feature is implemented via Java serialization mechanism, where a
"redo log" is saved during compi execution (in the ~/.compi folder).