Checks that pipeline has not changed in resume
This commit adds a check to stop resume if the pipeline file
content has changed. It uses a MD5 hash of the file. It also adds
a --flexible option to the resume command in order to skip this check.

This commit also refactors the resume implementation by not serializing
the Pipeline and Task objects, which is not needed.