Fixes bug when foreach iteration aborts affecting other iterations
When a foreach iteration aborts, the whole task is marked as
aborted, but it should not affect the other iterations of the
same loop, which should start anyway, because the loop concept assumes
that all iterations run in parallel and could start at the
same time (before any of the iterations abort).

This bug was detected when the thread pool is small (e.g. size 1),
where iterations do not run in parallel, but it should behave as
they where launched at the same time with an infinite thread pool.