Adds custom entrypoint.sh to the Compi Docker images
This commit adds an entrypoint.sh as entrypoint of the Compi
DK generated Docker images. This entrypoint makes easier to invoque
the "resume" functionality of Compi, as it detects the word
"resume" as the first parameter.

Moreover, this commit ensures that the pipeline.xml inside the
image is renamed by adding a suffix that depends on the pipeline's
name and version. This avoids the collision of pipeline file path
("/pipeline.xml") in all Compi-based docker images, which,
if all were run over the same shared volume to save the execution logs,
Compi will raise an error when trying to find the execution log of ç
a given pipeline.