Removes iteration notifications from "dummy iterations" from emtpy foreachs
Until now, for empty (no iteration) or skipped foreachs, a dummy iteration
process was submitted to the thread pool in order to avoid compi from
hanging waiting for the foreach task to finish. This dummy iterations
generated a message of start/finish in the CLI.

This commit fixes this by doing two things:

- Only generate a dummy iteration only for skipped foreachs, not for
empty foreachs. Skipped foreachs, as well, as all skipped tasks,
do not generate CLI messages of start/finish, so the problem of
CLI messages is fixed with this.

- Avoiding the main thread to block if there are no more tasks left to
run. During the fixation of this bug, we have realized that empty
foreachs were a problem only when they are the only tasks to be run.
This was because since the introduction of iteration-binded foreachs,
when a foreach starts, the dependant tasks do not further depend on the whole
foreach, but on its iterations. So, when an empty foreach starts,
the dependant tasks can also start since they do not depend on any
(inexistant iteration) and they will not be stalled forever. The only problem
was for pipelines with an only emtpy foreach. In this case, when Compi
starts this empty foreach, there are no more tasks left to run, and if
Compi waits in the syncMonitor, it will hang. This fix adds the condition
in order to avoid this unnecessary lock.